Add tests (+ fix bugs) & missing javadoc
